Add Cucumber test suite (bundle update required).
This introduces a new test suite that can be kicked off with:
  $ rake cucumber

There is one test that walks through creating a customer record,
plus all of the scaffolding to support it.

New dependencies need to be installed, and see the change to
database.yml.example, which you may have to apply to your database.yml
config file.