Task #420: Add name and phone number to user.
Changes the user list to show a display name for all users. The name is
a link to a new edit form that allows you to edit name, phone number, and
email address.  New users will be required to add a first name and last name,
but phone number is optional.

Updated test factories to ensure users have names as required.