Improve resource management interface.
Add active flag and split resources into two lists by active/inactive.
Add a quick toggle link to activate or deactivate a resource, and get
rid of the destroy action.  (If desired we can add it back.)

Update the create/edit actions to just return to the resources page.
The "show" page isn't a terribly nice page to land on.