[#678] Tighten User Authentication
Removing the Devise rememberable module from the User model and removing the
associated attributes from the DB, thus getting rid of the "Remember Me"
feature on the login page.

Also making use of the validatable module to enforce a minimum password length
of 8 characters and adding a custom validation to enforce a strong password
format of "at least one number and at least one non-alphanumeric character".
NOTE that this means all test users are now generated with a password of
"password 1" (p-a-s-s-w-o-r-d-[space]-1). Existing users will not be forced to
change their password nor will it be done automatically, but they will be
forced to use the new format if they choose to reset their password.