Migrate app to NUBIC surveyor 0.22.0.
This removes the dependency on the openplans surveyor fork, which has not been
actively maintained and will eventually cause issues as Rails moves forward.

Several migrations were added to accomodate the new fields required by
surveyor.  Fixed a couple code incompatibilities and a display_order issue
(surveyor apparently requires display_order to be set, otherwise the order is
not preserved from the JSON).