Upgrade cucumber and fix test issue.
I believe capybara is normalizing the spaces in text before performing the
search in have_content, so it was not finding the string "  1-May-2012"
because of the two leading spaces.

I'm still seeing a surveyor test fail because the sections are appearing in
the wrong order.  I believe this may be a bug that will be addressed by
getting back to the canonical surveyor repository:

   https://github.com/NUBIC/surveyor/issues/246